refact: move account to wallet package
