--- packages/account/README.md
@@ -1,53 +1,0 @@
-# `@fuel-ts/account`
-
-**@fuel-ts/account** is a sub-module for interacting with **Fuel**.
-
-This module contains a class that implements basic Account functionality, including querying the balance of an account and transactions.
-
-# Table of contents
-
-- [`@fuel-ts/account`](#fuel-tsaccount)
-- [Table of contents](#table-of-contents)
-  - [Documentation](#documentation)
-  - [Usage](#usage)
-    - [Installation](#installation)
-    - [Full SDK Installation](#full-sdk-installation)
-  - [Contributing](#contributing)
-  - [Changelog](#changelog)
-  - [License](#license)
-
-## Documentation
-
-See [Fuel-ts Documentation](https://fuellabs.github.io/fuels-ts/packages/fuel-ts-account/)
-
-## Usage
-
-### Installation
-
-```sh
-yarn add @fuel-ts/account
-# or
-npm add @fuel-ts/account
-```
-
-### Full SDK Installation
-
-Alternatively, we recommend you install the [complete SDK](https://github.com/FuelLabs/fuels-ts) using the umbrella package:
-
-```sh
-yarn add fuels
-# or
-npm add fuels
-```
-
-## Contributing
-
-In order to contribute to `@fuel-ts/account`, please see the main [fuels-ts](https://github.com/FuelLabs/fuels-ts) monorepo.
-
-## Changelog
-
-The `@fuel-ts/account` changelog can be found at [CHANGELOG](./CHANGELOG.md).
-
-## License
-
-The primary license for `@fuel-ts/account` is `Apache 2.0`, see [LICENSE](./LICENSE).
