--- packages/account/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@fuel-ts/account",
-  "version": "0.31.0",
-  "description": "Base Account Class for Fuel",
-  "author": "Fuel Labs <contact@fuel.sh> (https://fuel.network/)",
-  "typedoc": {
-    "entryPoint": "./src/index.ts"
-  },
-  "main": "src/index.ts",
-  "exports": {
-    ".": {
-      "require": "./dist/index.js",
-      "default": "./dist/index.mjs"
-    },
-    "./test-utils": {
-      "require": "./dist/test-utils.js",
-      "default": "./dist/test-utils.mjs"
-    }
-  },
-  "publishConfig": {
-    "main": "dist/index.js",
-    "module": "dist/index.mjs",
-    "types": "dist/index.d.ts",
-    "typings": "dist/index.d.ts"
-  },
-  "files": [
-    "dist"
-  ],
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@ethersproject/bytes": "^5.7.0",
-    "@fuel-ts/address": "workspace:*",
-    "@fuel-ts/constants": "workspace:*",
-    "@fuel-ts/interfaces": "workspace:*",
-    "@fuel-ts/math": "workspace:*",
-    "@fuel-ts/providers": "workspace:*",
-    "@fuel-ts/transactions": "workspace:*"
-  },
-  "scripts": {
-    "build": "tsup --dts"
-  }
-}
