--- packages/account/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "downlevelIteration": true,
-    "outDir": "./dist",
-    "rootDir": "./",
-    "baseUrl": "./"
-  },
-  "include": ["src", "test"]
-}
