+++ packages/predicate/package.json
@@ -30,7 +30,7 @@
     "@fuel-ts/address": "workspace:*",
     "@fuel-ts/interfaces": "workspace:*",
     "@fuel-ts/versions": "workspace:*",
-    "@fuel-ts/account": "workspace:*",
+    "@fuel-ts/wallet": "workspace:*",
     "@fuel-ts/providers": "workspace:*",
     "@fuel-ts/transactions": "workspace:*",
     "@fuel-ts/merkle": "workspace:*"
