+++ packages/predicate/src/predicate.test.ts
@@ -1,10 +1,10 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { hexlify } from '@ethersproject/bytes';
-import { Account } from '@fuel-ts/account';
 import { Address } from '@fuel-ts/address';
 import { bn } from '@fuel-ts/math';
 import { ScriptTransactionRequest, CoinStatus } from '@fuel-ts/providers';
 import type { InputCoin } from '@fuel-ts/transactions';
+import { Account } from '@fuel-ts/wallet';
 
 import { Predicate } from './predicate';
