+++ packages/wallet/package.json
@@ -39,9 +39,7 @@
     "@fuel-ts/mnemonic": "workspace:*",
     "@fuel-ts/providers": "workspace:*",
     "@fuel-ts/signer": "workspace:*",
-    "@fuel-ts/transactions": "workspace:*",
-    "@fuel-ts/predicate": "workspace:*",
-    "@fuel-ts/account": "workspace:*"
+    "@fuel-ts/transactions": "workspace:*"
   },
   "devDependencies": {
     "@fuel-ts/address": "workspace:*",
