+++ packages/wallet/src/index.ts
@@ -1,3 +1,4 @@
 export * from './base-unlocked-wallet';
+export * from './account';
 export * from './wallet';
 export * from './wallets';
