+++ packages/wallet/src/wallets.ts
@@ -4,12 +4,12 @@
 //
 /* eslint-disable max-classes-per-file */
 import type { BytesLike } from '@ethersproject/bytes';
-import { Account } from '@fuel-ts/account';
 import { HDWallet } from '@fuel-ts/hdwallet';
 import { Mnemonic } from '@fuel-ts/mnemonic';
 import type { Provider } from '@fuel-ts/providers';
 import { Signer } from '@fuel-ts/signer';
 
+import { Account } from './account';
 import { BaseWalletUnlocked } from './base-unlocked-wallet';
 import type { GenerateOptions } from './types/GenerateOptions';
